requirements.txt
playwright>=1.38
pandas>=2.0
openpyxl>=3.1
psycopg2-binary>=2.9
cassandra-driver>=3.27
